% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jellyfish.R
\name{Jellyfish}
\alias{Jellyfish}
\title{Jellyfish}
\usage{
Jellyfish(jellyfish)
}
\arguments{
\item{jellyfish}{}
}
\value{
number of jellyfish
}
\description{
Jellyfish
}
\examples{
give it a number of jellyfish
}
