% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{Calculate my snacks}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grapes in grams}

\item{apples}{number of apples}

\item{burritos}{how many burritos}
}
\value{
what to have for a snack
}
\description{
Calculate my snacks
}
\examples{
snacks(grapes = 2.5, apples = 6, burritos = 1)
}
