% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripple_that.R
\name{triple_that}
\alias{triple_that}
\title{Triples anything}
\usage{
triple_that(number)
}
\arguments{
\item{number}{}
}
\value{
three times your input
}
\description{
Triples anything
}
\examples{
input a numeric value
}
