% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekend_fun.R
\name{weekend_fun}
\alias{weekend_fun}
\title{Weekend Fun}
\usage{
weekend_fun(snacks, friends, activities, rain)
}
\arguments{
\item{rain}{}
}
\value{
how much fun you will have this weekend
}
\description{
Weekend Fun
}
\examples{
use to calculate how much fun you will have in units of snacks, friends, rain, and activities
}
